chess-expert-system/src/rules/missed_checkmate.py
import chess

class MissedCheckmateRule:
    def __init__(self, game):
        self.game = game

    def evaluate(self):
        # Analyze the game state to identify missed checkmate opportunities
        for move in self.game.mainline_moves():
            if self.is_checkmate_opportunity(move):
                print(f"Missed checkmate opportunity after move: {move}")

    def is_checkmate_opportunity(self, move):
        # Logic to determine if a checkmate could have been delivered
        board = chess.Board()
        for m in self.game.mainline_moves():
            board.push(m)
            if m == move:
                break

        # Check if the current player has a checkmate move
        for candidate_move in board.legal_moves:
            board.push(candidate_move)
            if board.is_checkmate():
                board.pop()
                return True
            board.pop()

        return False