chess-expert-system/requirements.txt
experta
python-chess
pgn-parser